requirements.txt
streamlit
langchain
langchain-community
openai
python-dotenv
manim
numpy
